feat(app): started ui design
