+++ package.json
@@ -53,7 +53,6 @@
     ],
     "scripts": {
         "start": "pnpm run prebuild && tsdx watch",
-        "prebuild": "node -p \"'export const LIB_VERSION = ' + JSON.stringify(require('./package.json').version) + ';'\" > src/version.ts",
         "build:browser": "tsdx build --entry ./src/index.browser.ts --tsconfig ./tsconfig.browser.json --format esm,cjs --name browser",
         "build:node": "tsdx build --entry ./src/index.node.ts --tsconfig ./tsconfig.node.json --format cjs --name node",
         "build": "node scripts/build.js",
@@ -137,10 +136,8 @@
         "@azure/identity": "^4.5.0",
         "@azure/msal-browser": "^3.5.0",
         "@azure/storage-blob": "^12.25.0",
-        "@emotion/styled": "^11.11.0",
         "@google-cloud/storage": "^7.14.0",
         "@microsoft/microsoft-graph-client": "^3.0.7",
-        "@mui/material": "^5.14.18",
         "@types/gapi": "^0.0.47",
         "babel-plugin-module-resolver": "^5.0.0",
         "framer-motion": "^10.16.5",
@@ -153,13 +150,8 @@
         "uuid": "^9.0.1"
     },
     "peerDependencies": {
-        "@emotion/react": "^11.11.1",
-        "@emotion/styled": "^11.x",
-        "@mui/material": "^5.x",
-        "@mui/system": "^5.x",
         "react": "^18.2.0",
-        "react-dom": "^18.2.0",
-        "react-is": "^18.2.0"
+        "react-dom": "^18.2.0"
     },
     "pnpm": {
         "overrides": {
