+++ postcss.config.js
@@ -1,5 +1,6 @@
 module.exports = {
     plugins: {
+        'postcss-import': {},
         tailwindcss: {},
         autoprefixer: {},
     },
