+++ src/frontend/components/AdapterSelector.tsx
@@ -1,76 +1,128 @@
-import React from 'react'
+import { motion } from 'framer-motion'
+import React, { ChangeEventHandler, Dispatch, SetStateAction } from 'react'
 import { UploadAdapter } from '../../shared/types'
 import { useRootContext } from '../context/RootContext'
 import { uploadAdapterObject } from '../lib/constants'
 import ShouldRender from './shared/ShouldRender'
 
-const getH1ClassName = (mini: boolean) => {
-    let className = 'text-center dark:text-white '
-    className += !mini ? 'md:text-2xl' : ''
-
-    return className
+type Props = {
+    isDragging: boolean
+    setIsDragging: Dispatch<SetStateAction<boolean>>
 }
 
-export default function AdapterSelector() {
+export default function AdapterSelector({ isDragging, setIsDragging }: Props) {
     const {
         inputRef,
         setView,
-        props: { onIntegrationClick, uploadAdapters, mini },
+        setFiles,
+        props: {
+            accept,
+            multiple,
+            onIntegrationClick,
+            uploadAdapters,
+            mini,
+            maxFileSize,
+            onFileDragLeave,
+            onFileDragOver,
+            onFileDrop,
+        },
     } = useRootContext()
     const chosenAdapters = Object.values(uploadAdapterObject).filter(item =>
         uploadAdapters.includes(item.id),
     )
     const handleAdapterClick = (adapterId: UploadAdapter) => {
         onIntegrationClick(adapterId)
-        if (adapterId === UploadAdapter.INTERNAL)
-            inputRef && inputRef?.current?.click()
+        if (adapterId === UploadAdapter.INTERNAL) inputRef.current?.click()
         else setView(adapterId)
     }
 
+    const handleInputFileChange: ChangeEventHandler<HTMLInputElement> = e => {
+        setFiles(Array.from(e.currentTarget.files || []))
+        e.currentTarget.value = ''
+    }
+
     return (
-        <div
-            className={
-                !mini
-                    ? 'flex h-full w-full flex-col items-center justify-center gap-6'
-                    : ''
-            }
+        <motion.div
+            initial={{ opacity: 0 }}
+            animate={{ opacity: 1 }}
+            exit={{ opacity: 0 }}
+            className={`flex flex-col items-center justify-center gap-10 rounded-lg border border-[#1849D6] p-[30px] dark:border-[#30C5F7] lg:gap-14 ${
+                isDragging
+                    ? 'bg-[#E7ECFC] backdrop-blur-sm dark:bg-[#045671]'
+                    : 'border-dashed'
+            }`}
+            onDragOver={e => {
+                e.preventDefault()
+                setIsDragging(true)
+                e.dataTransfer.dropEffect = 'copy'
+
+                const files = Array.from(e.dataTransfer.files)
+                onFileDragOver(files)
+            }}
+            onDragLeave={e => {
+                e.preventDefault()
+                setIsDragging(false)
+
+                const files = Array.from(e.dataTransfer.files)
+                onFileDragLeave(files)
+            }}
+            onDrop={e => {
+                e.preventDefault()
+
+                const droppedFiles = Array.from(e.dataTransfer.files)
+
+                onFileDrop(droppedFiles)
+                setFiles(droppedFiles)
+
+                setIsDragging(false)
+            }}
         >
-            <h1 className={getH1ClassName(mini)}>
-                Drop your files here or{' '}
-                <button
-                    className="text-[#3782da] hover:underline"
-                    onClick={() => inputRef && inputRef.current!.click()}
-                    type="button"
-                >
-                    Click to browse
-                </button>{' '}
-                {!mini && <span>or import from:</span>}
-            </h1>
             <ShouldRender if={!mini}>
-                <div className="grid grid-cols-3 grid-rows-2 sm:grid-cols-4 md:grid-cols-6">
+                <div className="flex flex-wrap items-center justify-center gap-[26px] lg:gap-[30px]">
                     {chosenAdapters.map(({ Icon, id, name }) => (
                         <button
                             type="button"
                             key={id}
-                            className="group relative mb-4 flex flex-col items-center justify-center gap-1 rounded-md p-2 px-4 text-sm transition-all duration-300 hover:bg-[#e9ecef] active:bg-[#dfe6f1] disabled:pointer-events-none disabled:opacity-30 dark:hover:bg-[#282828] dark:active:bg-[#333]"
+                            className="group flex flex-col items-center justify-center gap-[3px] rounded-lg lg:gap-[6px]"
                             onKeyDown={e => {
                                 if (e.key === 'Enter') e.preventDefault()
                             }}
                             onClick={() => handleAdapterClick(id)}
                         >
-                            <span className="rounded-lg bg-white p-[6px] text-2xl shadow dark:bg-[#323232]">
+                            <span className="rounded-lg bg-white p-[6px] text-2xl font-semibold shadow group-hover:scale-110 dark:bg-[#323232]">
                                 <Icon />
                             </span>
-                            <span className="text-[#525252] dark:text-[#777]">
+                            <span className=" text-xs text-[#242634] dark:text-[#6D6D6D]">
                                 {name}
-                            </span>
-                            <span className="absolute -bottom-2 hidden opacity-50 group-disabled:block">
-                                (soon)
                             </span>
                         </button>
                     ))}
                 </div>
             </ShouldRender>
-        </div>
+            <input
+                type="file"
+                accept={accept}
+                className="hidden"
+                ref={inputRef}
+                multiple={mini ? false : multiple}
+                onChange={handleInputFileChange}
+            />
+            <div className="flex flex-col gap-2">
+                <div className="flex items-center gap-1">
+                    <span className="text-xs text-[#0B0B0B] dark:text-white lg:text-sm">
+                        Drag your file(s) or
+                    </span>
+                    <span
+                        className="cursor-pointer text-xs font-semibold text-[#0E2ADD] dark:text-[#59D1F9] lg:text-sm"
+                        onClick={() => inputRef.current?.click()}
+                    >
+                        browse
+                    </span>
+                </div>
+                <p className="text-xs text-[#6D6D6D] lg:text-sm">
+                    Max {maxFileSize.size} {maxFileSize.unit} files are allowed
+                </p>
+            </div>
+        </motion.div>
     )
 }
