--- src/version.ts
@@ -1,1 +1,0 @@
-export const LIB_VERSION = "2.4.0";
