+++ stories/LocalDigitalOceanUpload.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import React, { useRef } from 'react'
+import React from 'react'
 import UpupUploader from '../src/frontend/UpupUploader'
-import { UpupProvider, UpupUploaderRef } from '../src/shared/types'
+import { UpupProvider } from '../src/shared/types'
 
 const meta = {
     title: 'Cloud Storage/Local to DigitalOcean Upload',
@@ -20,35 +20,14 @@ const meta = {
 export default meta
 type Story = StoryObj<typeof UpupUploader>
 
-const LocalUploader = () => {
-    const upupRef = useRef<UpupUploaderRef>(null)
-
-    const handleUpload = async () => {
-        try {
-            await upupRef.current?.uploadFiles()
-        } catch (error) {
-            console.error('Error uploading selected files:', error)
-        }
-    }
-
-    return (
-        <div className="flex flex-col gap-4">
-            <UpupUploader
-                ref={upupRef}
-                provider={UpupProvider.DigitalOcean}
-                tokenEndpoint="http://localhost:3001/api/storage/digitalocean/upload-url"
-                accept="*"
-                maxFileSize={{ size: 10, unit: 'MB' }}
-            />
-            <button
-                className="rounded bg-blue-500 px-4 py-2 text-white hover:bg-blue-600"
-                onClick={handleUpload}
-            >
-                Upload to DigitalOcean
-            </button>
-        </div>
-    )
-}
+const LocalUploader = () => (
+    <UpupUploader
+        provider={UpupProvider.DigitalOcean}
+        tokenEndpoint="http://localhost:3001/api/storage/digitalocean/upload-url"
+        accept="*"
+        maxFileSize={{ size: 10, unit: 'MB' }}
+    />
+)
 
 export const Default: Story = {
     render: () => <LocalUploader />,
