--- todo.md
@@ -1,19 +1,0 @@
-# todo
-
-## CURRENT
-
--   useProgress
-
--   Handle chunked uploads for large files
--   Add support for resumeable uploads
--   Add retry logic
--   Improved error handling
-
-## IMPROVEMENTS: Later
-
--   Component props for proper error and info updates on ongoing processes: onError, onUploading, etc
-
-## DOCUMENTATION
-
--   IDE documentation for the component
--   unit tests
